Added example of twitch api
Added a small example site to the repository that is aimed to describe,
in depth, the processes of utilizing the Twitch API. It will be heavily
documented and upon full understanding of the API the scripts will be
implemented into the actual site. For now this serves as an easy way to
keep everyone up to speed on the API as I'm learning it.